feat: Added custom widgets theme impl
